Removed all ABS code and replaced it untested
